moved internal code from ortholotree.py into internal.py
all code previously in ortholotree is undergoing major restructuring
for better comprehension of functions and their relationships to
external dependencies, and for better incorporation of test code.

The ortholotree.py was therefore moved and renamed (in hindsight,
perhaps this was unnecessarily confusing), as the bulk of the code
was to remain in this file, but now labelled as the 'internal'
module.

All code calling external functions has been put into a wrapper
module called external, and the CLI functionality and API will be
in a script called ortholotree.py, which will access the internal
and external modules to execute.