updated alignment viewing function and phmmer parser
alignment viewer was split to allow the matplotlib object to be
sent to a different script for manipulation (such as a web server).
This addition is backwards compatible with the old scripts, as
all original variables still apply.

Also improved phmmer output file parsing.

Also added a gene counting function (which counts number of genes
provided in list and fasta format) for use in the hmmer build
functions. Hmmbuild will now automatically build a model if more
than 1 sequences (in either list or fasta format) are provided.