Added ability to view search fasta sequences in tree + alignment
Until now, any novel sequences that were used for searching/building
a hmm model for searching, would be discarded and only the identified
sequences in the selected species would be aligned and added to the
phylogeny. Now a temporary file is created that contains any
fasta sequences used in the search query, to allow phmmer to search
against the query sequences and return a score and p-value. The
sequences can also be added to the tree, allowing for analysis of
species that are not in the database.

Some additional species were also added to the default species list.