added visualization of consensus sequence in alignment
The previous alignment view showed the number of gaps within an
alignment and the position of large gaps, but the consensus between
positions from different sequences was not visible. This update
provides an additional bar on the bottom of the figure, allowing
a color-coded scale of alignment consensus as an additional measure
of the alignment quality.