improved regular expression searching of deflines
new regular expression allows for partial matching of NCBI
formatted entries, but still prevents partial matching of
substrings. ie, LOC12345 will match >gi|235498|ref|LOC12345| but it
will  not match >LOC1234.

The remove name dross is also more appropriately coded to
substitute all '|' for '\|', so that it will not be interpreted as
an RE 'OR' statement, but will only match a pipe present in the
search string.